wpt: removing test specific code for sensor timestamp request
during testing, mock-sensor is used, values and timestamp are
artificially generated in the test.

to comply with test restriction, IsRunningWebTests() has been
used to simulate timestamp.
IsRunningWebTests() is only working under CI wpt testing, and
does not work in upstream wpt.

This patch removes the test specific code from blink, and
changes the timestamp value to be acceptable for the calculation
of sensor.timestamp

Bug: 1135537
Change-Id: I94ca3e4ba63ce99fcc64bef070374483db218ebf